requirements.txt
asgiref==3.7.2
click==8.1.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
packaging==23.1
psycopg2==2.9.7
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
whitenoise==6.5.0
